requirements.txt
# TRON and Escrow dependencies
tronpy>=0.4.0
cryptography>=3.4.8
qrcode[pil]>=7.0

# Telegram Bot
python-telegram-bot>=20.0

# HTTP and async
requests>=2.32.4
aiohttp>=3.12.14
aiofiles>=23.0.0
aiohappyeyeballs>=2.6.1
aiosignal>=1.4.0

# Crypto API dependencies
python-binance>=1.0.29
pycoingecko>=3.1.0
yfinance>=0.2.0
pybit>=5.7.0

# Data handling
python-dateutil>=2.9.0.post0
dateparser>=1.2.2
pytz>=2025.2

# General utilities
attrs>=25.3.0
certifi>=2025.7.14
charset-normalizer>=3.4.2
frozenlist>=1.7.0
idna>=3.10
multidict>=6.6.3
propcache>=0.3.2
pycryptodome>=3.23.0
regex>=2024.11.6
six>=1.17.0
tzlocal>=5.3.1
urllib3>=2.5.0
websockets>=15.0.1
yarl>=1.20.1

# Environment variables
python-dotenv>=1.0.0